pyproject.toml
[project]
name = "card-flip-agent"
version = "0.1.0"
description = "AI agent for finding profitable card flips"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]


[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"